requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langgraph>=0.0.20

# Vector Store
faiss-cpu>=1.9.0

# Utilities
pydantic>=2.5.0
python-multipart>=0.0.6